% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{bin}
\alias{bin}
\title{Bin observations}
\usage{
bin(sp, n, n_bins = TRUE, summary = FALSE)
}
\arguments{
\item{sp}{data frame containing species observations as columns with an
environmental gradient as the first column (needs work).}

\item{n}{scalar value indicating the number of bins or number of
records per bin (see \code{n_bins}).}

\item{n_bins}{logical indicating if \code{n} is the number of bins
(\code{TURE}; default) or records per bin.}

\item{summary}{logical indicating if \code{bin} should return the binned
records (\code{FALSE}; default), or a summary bins}
}
\value{
If \code{summary = FALSE} (default), \code{bin} returns data frame
  containing the binned species records and environmental gradient.  Both of
  these are means. If presence/absence data are used (codes as 0/1), the
  mean is also a probability. If \code{summary = TRUE}, \code{bin} returns
  a table of the number of records per bin.
}
\description{
\code{bin} combines species observations based on input environmental
 gradient. Number of bins or number of records can be specified. Resulting
 bins can also be examined using the \code{summary} option.
}
\details{
\code{bin} is used to "bin" (i.e., combine by taking the mean) species
species observations for taxa used to create Biotic Response functions
(\code{\link{est_brc}}). Unlike other functions in package \code{iec},
function \code{bin} takes a data frame that contains the environmental
gradient as the first column.
}
\examples{
x <- data.frame(gradient = round(abs(rnorm(107)), 2),
                sp1 = rpois(107, 0.7),
                sp2 = rpois(107, 0.5))

bin(x, n = 7, summary = TRUE)
x_bin <- bin(x, n = 7)
}

