% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{fish_grad}
\alias{fish_grad}
\title{Environmental gradient for fish sites}
\format{\code{fish_grad} is a single column data frame containing
  environmental gradient scores from 25 sites (rows).  Site corresponds to
  sites (rows) in \code{\link{fish_sp}}.}
\usage{
fish_grad
}
\description{
Environmental gradient scores for sites in \code{\link{fish_sp}}. Larger
values indicate less desirable sites, so the scale will need to be inverted
for use with \code{\link{est_brc}}.
}
\examples{
data(fish_grad)
grad <- scale10(fish_grad, invert = TRUE)
# grad is now scaled to 0-10 with 10 being the least degraded.
summary(grad)
}
\seealso{
\code{\link{scale10}}, \code{\link{est_brc}}
}
\keyword{datasets}

