% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{fish_sp}
\alias{fish_sp}
\title{Fish observations}
\format{\code{fish_sp} is a data frame containing records of 4 fish species
  (columns) from 25 sites (rows).}
\usage{
fish_sp
}
\description{
Small dataset of fish observations.
}
\seealso{
\code{\link{fish_grad}}
}
\keyword{datasets}

