% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_sens}
\alias{get_sens}
\title{Sensitivity of taxa to an environmental gradient.}
\usage{
get_sens(sp, env_grad, mid = 5)
}
\arguments{
\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{env_grad}{numeric vector of environmental gradient scores, one for each
site.}

\item{mid}{numeric scalar taken as the gradient's midpoint for sensitivity
(default is 5).}
}
\value{
A data frame containing each taxon's sensitivity (\code{Sens}) to the
  gradient, and the number of sites the they were detected at (\code{n}).
}
\description{
\code{get_sens} returns the sensitivity of each taxon in a community data
frame (\code{sp}) to an environmental gradient (\code{env_grad}).
}
\details{
Sensitivity is defined here by first finding how many detections were made
above and below the midpoint (\code{mid}) of the gradient, taking their
difference, and finally dividing this by the number of detections. Taxa with
more detections below (\code{mid}) than above will have a negative
sensitivity and vice versa. Note that while function \code{\link{est_brc}}
requires that a reference gradient scaled to 0-10, \code{get_sens} does not
require this.
}
\examples{
x <- function() sample(0:1, 10, replace = TRUE)
sp <- data.frame(sp1 = x(), sp2 = x(), sp3 = x())
grad <- runif(10, 0, 10)
get_sens(sp, grad)
}

