% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{nlr2}
\alias{nlr2}
\title{Nonlinear R-squared}
\usage{
nlr2(observed, expected)
}
\arguments{
\item{observed}{a numeric vector of observations.}

\item{expected}{a numeric vector of expected values.}
}
\value{
The nonlinear R-squared value (numeric scalar).
}
\description{
\code{nlr2} calculates a nonlinear R-squared for the observed values.
}
\details{
This function calculates the non-linear R-squared.  The equation used is:
\deqn{R^{2} = 1 - \frac{SS_{reg}}{SS_{tot}}}{R^2 = 1 - (SS_reg / SS_tot)}
(see p. 34-35
\href{http://www.mcb5068.wustl.edu/MCB/Lecturers/Baranski/Articles/RegressionBook.pdf}{here}).
Note that this nonlinear R-squared is not anything squared so it can result
a negative value. Negative values indicate poor model fit.
}

