% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_resp}
\alias{plot_resp}
\title{Plot response surface for sites.}
\usage{
plot_resp(sp, iec, brc, method = "pa")
}
\arguments{
\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{iec}{the output data frame from \code{\link{est_iec}}.}

\item{brc}{BRC data frame generated by \code{\link{est_brc}}.}

\item{method}{one of \code{"pa"} (default; CalcPA) or \code{"q"} (CalcQ).}
}
\description{
\code{plot_resp} plots the response surface for the IEC score at each site in
data frame \code{sp}.
}
\details{
Plots are generated for each site in \code{sp} with probability on the y-axis
and IEC score on the x-axis. The "best" value found by \code{\link{est_iec}}
is also plotted as a vertical line.  Note that \code{sp}, \code{brc}, and
\code{method} must be the same as those used to generate \code{iec} using
\code{\link{est_iec}}. Also, note that this function can be quite time
consuming. It is generally best to save the output to an external plotting
device (e.g., \code{\link[grDevices]{pdf}})
}
\seealso{
\code{\link{est_iec}}, \code{\link[grDevices]{Devices}}
}

