% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale10}
\alias{scale10}
\title{Scale environmental gradient to 0-10}
\usage{
scale10(env_grad, invert = FALSE)
}
\arguments{
\item{env_grad}{numeric vector of environmental gradient scores, one for each
site.}

\item{invert}{logical indicating if \code{env_grad} should be inverted
(default is \code{FALSE}).}
}
\value{
A numeric vector with \code{env_grad} scaled to 0-10.
}
\description{
\code{scale10} scales an environmental gradient to the range 0-10.
Optionally, it can invert the scale.  This is essential for use with
\code{\link{est_brc}}.
}
\details{
\code{scale10} is used to scale an environmental gradient for use in
generating Biotic Response Curves with \code{\link{est_brc}}.  The reference
gradient returned by \code{scale10} has minimum 0 and maximum 10. If the
input environmental gradient has larger values for less desirable sites, use
\code{invert = TRUE} to invert the gradient so that 10 represents the most
desirable condition.
}
\examples{
grad <- runif(10)
scale10(grad, TRUE)
}

