% Generated by roxygen2 (4.0.2): do not edit by hand
\name{brc_pdf}
\alias{brc_pdf}
\title{Save BRC plots to PDF file.}
\usage{
brc_pdf(out_pdf, sp, brc, ref_grad, min_lof = 1)
}
\arguments{
\item{out_pdf}{name of output PDF file.}

\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{brc}{BRC data frame generated by \code{\link{est_brc}}}

\item{ref_grad}{numeric vector of reference gradient scores (0-10), one for
each site.}

\item{min_lof}{numeric value indicating minimum LOF to label.}
}
\description{
\code{brc_pdf} saves the output from \code{\link{plot_lof}} and
\code{\link{plot_brc}} to a PDF.
}
\details{
\code{brc_pdf} checks for if the output file already exists and returns an
error is it does.  Otherwise, \code{brc_pdf} opens a PDF device to print to
(\code{out_pdf}) and then saves the output from \code{\link{plot_lof}} and
\code{\link{plot_brc}} to the PDF file.  This is done within
\code{\link[base]{tryCatch}} such that \code{\link[grDevices]{dev.off}} is
safely called even if the printing fails.
}
\examples{
brc_pdf("output.pdf", sp, brc, grad)
}
\seealso{
\code{\link{est_brc}}, \code{\link{scale10}},
  \code{\link[grDevices]{Devices}}
}

