% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est_brc}
\alias{est_brc}
\title{Estimate Biotic Response Curves (BRC).}
\usage{
est_brc(sp, ref_grad)
}
\arguments{
\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{ref_grad}{numeric vector of reference gradient scores (0-10), one for
each site.}
}
\value{
A data frame defining BRCs for each species or other taxa in
  \code{sp}.
}
\description{
\code{est_brc} generates biotic response curves (BRC) for each species or
taxon in \code{sp} in relation to environmental reference gradient
\code{ref_grad}.
}
\details{
The Biotic Response Curves or functions returned by \code{est_brc} are normal
curves fit to the observations in \code{sp} using a lack-of-fit (LOF)
criteria. BRCs consist of a normal curve defined by mean (mu) and standard
deviation (sigma), which is multiplied (scaled) a height factor (H).  The
reference gradient (\code{ref_grad}) input to \code{est_brc} must be a
numeric vector scaled to 0-10 where 10 represents the least impacted site.
Use \code{\link{scale10}} to scale the reference gradient if needed. Note
that \code{ref_grad} must have the same order by site as \code{sp}. The
results of \code{est_brc} are used to give sites an Index of Ecological
Condition score using function \code{\link{est_iec}}.
}
\references{
Gnass Giese, E.E., R.W. Howe, A.T. Wolf, N.A. Miller, and N.G.
  Walton. 2014. Sensitivity of breeding birds to the "human footprint" in
  western Great Lakes forest landscapes. In review.

  Howe, R.W., R. R. Regal, J.M. Hanowski, G.J. Niemi, N.P. Danz, and C.R.
  Smith.  2007a.  An index of ecological condition based on bird assemblages
  in Great Lakes coastal wetlands.  Journal of Great Lakes Research 33
  (Special Issue 3): 93-105.

  Howe, R.W., R. R. Regal, G.J. Niemi, N.P. Danz, J.M. Hanowski. 2007b.  A
  probability-based indicator of ecological condition. Ecological Indicators
  7:793-806.
}
\seealso{
\code{\link{scale10}}
}

