% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est_iec}
\alias{est_iec}
\title{Estimate IEC site scores.}
\usage{
est_iec(sp, brc, method = "pa", n_reps = 30)
}
\arguments{
\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{brc}{BRC data frame generated by \code{\link{est_brc}}}

\item{method}{one of \code{"pa"} (default; CalcPA) or \code{"q"} (CalcQ).}

\item{n_reps}{numeric scalar setting the number of random starts used in
optimization (default is 30).}
}
\value{
A data frame containing the IEC scores for each site in \code{sp} and
  the corresponding likelihood of the score.
}
\description{
\code{est_iec} estimates the Index of Ecological Condition (IEC) for each
site in \code{sp} based on the biotic response curves (BRC) in \code{brc}.
}
\details{
An Index of Ecological Condition (IEC) score is estimated for each site in
data frame \code{sp}. The default method treats observations as
presence/absence (\code{method = "pa"}).  When quantitative observations are
available, set \code{method} to \code{q}. \code{n_reps} determines the number
of times the optimization function (\code{\link[stats]{nlminb}}) is run with
different starting values. The column names in \code{sp} and column 1 of
\code{brc} must contain the same taxa in the same order. \code{n_reps} can be
set to any positive integer, but the authors recommend against using values
of less than 10 except for testing purposes.  Larger values will take longer
to process.  The default value is 30.
}
\references{
Gnass Giese, E.E., R.W. Howe, A.T. Wolf, N.A. Miller, and N.G.
  Walton. 2014. Sensitivity of breeding birds to the "human footprint" in
  western Great Lakes forest landscapes. In review.

  Howe, R.W., R. R. Regal, J.M. Hanowski, G.J. Niemi, N.P. Danz, and C.R.
  Smith.  2007a.  An index of ecological condition based on bird assemblages
  in Great Lakes coastal wetlands.  Journal of Great Lakes Research 33
  (Special Issue 3): 93-105.

  Howe, R.W., R. R. Regal, G.J. Niemi, N.P. Danz, J.M. Hanowski. 2007b.  A
  probability-based indicator of ecological condition. Ecological Indicators
  7:793-806.
}
\seealso{
\code{\link{est_brc}} to generate biotic response curves.
}

