% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_brc}
\alias{plot_brc}
\title{Plot BRC}
\usage{
plot_brc(sp, brc, ref_grad)
}
\arguments{
\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{brc}{BRC data frame generated by \code{\link{est_brc}}}

\item{ref_grad}{numeric vector of reference gradient scores (0-10), one for
each site.}
}
\description{
\code{plot_brc} plots biotic response curves generated by
\code{\link{est_brc}} along with the observations used to generate them.
}
\details{
BRC plots are generated for each taxa in data frames \code{sp} and
\code{brc}.  \code{sp} and \code{brc} must contain the same taxa in the same
order.  Generally, \code{sp} and \code{ref_grad} will be the same variables
used to generate \code{brc} with \code{\link{est_brc}}.  Because many plots
will be generated by this function, it may be best to send it to a file
plotting device (e.g., \code{\link[grDevices]{pdf}}).
}
\examples{
plot_brc(sp, brc, grad)
}
\seealso{
\code{\link{est_brc}}, \code{\link{scale10}},
  \code{\link[grDevices]{Devices}}
}

