% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_brc}
\alias{plot_brc}
\title{Plot BRC}
\usage{
plot_brc(sp, brc, ref_grad, env_bad = NULL, env_good = NULL,
  ylab = "Response", sub_xlab = NULL, sub_main = NULL)
}
\arguments{
\item{sp}{community data frame (sites as rows, taxa as columns, observations
as values).}

\item{brc}{BRC data frame generated by \code{\link{est_brc}}.}

\item{ref_grad}{numeric vector of reference gradient scores (0-10), one for
each site.}

\item{env_bad}{numeric scalar worst condition of original environmental
gradient (optional).}

\item{env_good}{numeric scalar best condition of original environmental
gradient (optional).}

\item{ylab}{a title for the y axis (optional; defaults to "Response").}

\item{sub_xlab}{name of environmental gradient (optional);
sub title to x-axis.}

\item{sub_main}{a sub title for the plots (optional); printed below main
title (taxon name).}
}
\description{
\code{plot_brc} plots biotic response curves generated by
\code{\link{est_brc}} along with the observations used to generate them.
}
\details{
BRC plots are generated for each taxa in data frames \code{sp} and
\code{brc}.  \code{sp} and \code{brc} must contain the same taxa in the same
order.  Generally, \code{sp} and \code{ref_grad} will be the same variables
used to generate \code{brc} with \code{\link{est_brc}}.  Because many plots
will be generated by this function, it may be best to send it to a file
plotting device (e.g., \code{\link[grDevices]{pdf}}).

The optional arguments, \code{env_bad} and \code{env_good}, provide a way to
label each end of the x-axis with the worst and best observed values of the
original environmental gradient. This can be helpful while exploring the
data to see if a gradient has been scaled correctly (inverted or not).
\code{sub_xlab} can be used to add the original name of the gradient to the
x-axis below the x-axis label. \code{sub_main} provides an optional sub title
printed below the main title (name of taxon extracted from input \code{sp}).
This can be a good place to print additional information about the plot such
at the project or input dataset.
}
\seealso{
\code{\link{est_brc}}, \code{\link{scale10}},
  \code{\link[grDevices]{Devices}}
}

